requirements.txt
adbutils==2.8.5
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.1
deprecation==2.1.0
filelock==3.18.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
gradio_client==1.8.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
retry==0.9.2
rsa==4.9
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
websockets==15.0.1
